fix(src): fix regroup
PR #1108 fixed the issue to prevent duplicate regrouping when the tree
has already been regrouped.
However, this fix caused only the first subtree to be executed when
regrouping needed to be performed on multiple subtrees.
Specifically, the execution results of the following manifests became
different from before:

- manifests/examples/pipelines/scenario-3.yml
- manifests/examples/pipelines/pipeline-with-aggregate.yml

This PR improves the method for checking if regrouping has been
completed and fixes it to correctly detect when regrouping needs to be
executed on multiple subtrees.

Additionally, it also adds the output results for the above two
manifests that were removed in PR #1037.

Signed-off-by: Mitsuru Kariya <mitsuru.kariya@nttdata.com>